tugas_mandiri/requirements.txt
django==3.2.8
djangorestframework==3.13.1